♻️ Refactor: InvitedDashboard 서치바, 리스트에서 데이터 페칭
